Migrated extended api calls to cbapi.py; changed command line args to match; updated licensing;
