+++ client_apis/python/example/process_search.py
@@ -1,43 +1,43 @@
-import sys
-import struct
-import socket
+#!/usr/bin/env python
+#
+#The MIT License (MIT)
+#
+# Copyright (c) 2015 Bit9 + Carbon Black
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+# -----------------------------------------------------------------------------
+#  <Short Description>
+#
+#  <Long Description>
+#
+#  last updated 2015-06-28 by Ben Johnson bjohnson@bit9.com
+#
+
 import pprint
-import optparse 
+from cbapi.util.cli_helpers import main_helper
 
-# in the github repo, cbapi is not in the example directory
-sys.path.append('../src/cbapi')
-
-import cbapi 
-
-def build_cli_parser():
-    parser = optparse.OptionParser(usage="%prog [options]", description="Perform a process search")
-
-    # for each supported output type, add an option
-    #
-    parser.add_option("-c", "--cburl", action="store", default=None, dest="url",
-                      help="CB server's URL.  e.g., http://127.0.0.1 ")
-    parser.add_option("-a", "--apitoken", action="store", default=None, dest="token",
-                      help="API Token for Carbon Black server")
-    parser.add_option("-n", "--no-ssl-verify", action="store_false", default=True, dest="ssl_verify",
-                      help="Do not verify server SSL certificate.")
-    parser.add_option("-q", "--query", action="store", default=None, dest="query",
-                      help="process query e.g. hostname:foo and netconn:127.0.0.1")
-    return parser
-
-def main(argv):
-    parser = build_cli_parser()
-    opts, args = parser.parse_args(argv)
-    if not opts.url or not opts.token or opts.query is None:
-        print "Missing required param; run with --help for usage"
-        sys.exit(-1)
-
-    # build a cbapi object
-    #
-    cb = cbapi.CbApi(opts.url, token=opts.token, ssl_verify=opts.ssl_verify)
+def main(cb, args):
 
     # perform a single process search
     #
-    processes = cb.process_search(opts.query)
+    processes = cb.process_search(args.get('query'))
     
     print "%-20s : %s" % ('Displayed Results', len(processes['results']))
     print "%-20s : %s" % ('Total Results', processes['total_results'])
@@ -48,5 +48,7 @@ def main(argv):
     for process in processes['results']:
         pprint.pprint(process)
         print '\n'
+
 if __name__ == "__main__":
-    sys.exit(main(sys.argv[1:]))
+    required_arg = ("-q", "--query", "store", None, "query", "Process search query")
+    main_helper("Generic process search", main, custom_required=[required_arg])
